Expand lport to 32-bits to match OF-DPA, moving tunnel ports to 0x0001xxxx

+++ hw/net/rocker/rocker.h
@@ -31,7 +31,7 @@ static inline GCC_FMT_ATTR(1, 2) int DPRINTF(const char *fmt, ...)
 struct world;
 struct rocker;
 
-int rx_produce(struct world *world, uint16_t lport,
+int rx_produce(struct world *world, uint32_t lport,
                const struct iovec *iov, int iovcnt);
 void rocker_update_irq(struct rocker *r);
 void rocker_irq_status_append(struct rocker *r, uint32_t irq_status);
