+++ hw/net/rocker/rocker_world.h
@@ -29,7 +29,7 @@ enum rocker_world_type {
 
 typedef int (world_init)(struct world *world);
 typedef void (world_uninit)(struct world *world);
-typedef ssize_t (world_ig)(struct world *world, uint16_t lport,
+typedef ssize_t (world_ig)(struct world *world, uint32_t lport,
                            const struct iovec *iov, int iovcnt);
 typedef int (world_cmd)(struct world *world, struct rocker_tlv *cmd_info_tlv);
 
@@ -40,7 +40,7 @@ struct world_ops {
     world_cmd *cmd;
 };
 
-ssize_t world_ingress(struct world *world, uint16_t lport,
+ssize_t world_ingress(struct world *world, uint32_t lport,
                       const struct iovec *iov, int iovcnt);
 int world_do_cmd(struct world *world, struct rocker_tlv *cmd_info_tlv);
